requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pandas
python-dotenv
requests
mistralai
langchain-community
langchain-mistralai
faiss-cpu
ragas
datasets
